backend/requirements.txt
asgiref==3.4.1
click==8.0.3
Django==3.2.8
django-filter==21.1
djangorestframework==3.12.4
Markdown==3.3.4
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.0
python-dotenv==0.19.1
pytz==2021.3
regex==2021.10.21
sqlparse==0.4.2
tomli==1.2.1
typing-extensions==3.10.0.2
yapf==0.31.0
